Voice Call or Group Call - Flutter UI
There are 3 screens in this project , First, one for the audio call which has a user pic in the background, then for group call UI you can add as many friends as you want. Last one for dial call which includes all functionality like add more people or record or make a video call.